SERVER-95040 Add test coverage for checkMetadataConsistency interruptability (#28571)
This adds two tests to validate the interruptability of
`checkMetadataConsistency` via `killOp`:

* A FSM test that kills `checkMetadataConsistency` at random times and
validates that the command either completes successfully or is
interrupted. This test verifies correctness, i.e. that the operation
does not get stuck or fails with an unexpected error code.

* A standalone test using fail points that validates that when killed,
both the top-level `checkMetadataConsistency` command as well as the
children ` _shardsvrCheckMetadataConsistency` commands are effectively
interrupted. In particular, interruptability is checked:
1. for the top-level command while sub-commands are being sent to
shards.
2. for the shard sub-command, where participant commands are (serially)
sent to each shard for each database, which ensures that the entire
operation does not dispatch any further work - this is important because
the total amount of sub-commands dispatched if the operation is not
interrupted is `O(nShards * nDatabases)`.
3. for the shard sub-command, where it waits for the DDL lock, which is
expected to be the main choke point where the command will need to wait
if there are concurrent DDL operations.