SERVER-97004: Move validation documentation into its directory and document errors/warnings (#29281)
# Issue

[SERVER-97004](https://jira.mongodb.org/browse/SERVER-97004)

# Description

To aid developers and improve discoverability of our documentation, we
move the validation-relevant documentation into the validate directory.
Note that this does not include "repair", which at time of writing still
lives partially in the collection directory.

Now that the documentation is in the expected location, we extend it
with a new section describing the decision process for choosing errors
or warnings when writing validation passes.