SERVER-97291 Split index bounds building into a separate library (#29309)
The motivation for this change is to allow CBR's cardinality estimator
to invoke `IndexBoundsBuilder::translate()` to be able to generate
bounds for certain MatchExpressions. But this is currently impossible
because `index_bounds_builder.cpp` is defined in the `query_planner`
library which depends on the `cardinality_estimator`, so we can't add a
dependency from the latter onto the former without introducing a cycle.
Thus, we have to refactor the functionality we need into a separate
library which both can depend on.

`index_bounds_builder.h/cpp` had several dependencies on files
previously defined in the `query_planner` target. For each one, we
detail how we broke the dependency.

1. `expression_geo_index_mapping.h/cpp` Decided to include this in the
new `index_bounds_builder` library. As they are logically in the same
"module" of index bounds building.
2. `interval_evaluation_tree.h/cpp` Same as (1).
3. `planner_ixselect.h/cpp`. This was for the following symbols:
    * `nodeIsSupportedByHashedIndex`: Move to `indexability.h`
    * `canUseIndexForNin`: Move to `indexability.h`
4. `planner_wildcard_helpers.h/cpp`. This was for the following symbols:
    * `getWildcardField`: Move to be member function of `IndexEntry`
* `translateWildcardIndexBoundsAndTightness`: Move the implementation to
`index_bounds_builder.cpp`. This introduced a new problem. This function
referenced the following symbols defined in `planner_wildcard_helpers`:
* `boundsOverlapObjectTypeBracket`: Move implementation to
`index_bounds.cpp` and removed its usage of
`IndexBoundsBuilder::makeRangeInterval` and construct an `Interval` by
hand.
* `findArrayIndexPathComponents`: Move to a new file
`index_multikey_helpers.h` and new target. This is required as both
`index_bounds_builder.cpp` and `planner_wildcard_helpers.cpp` needed to
use this function.