SERVER-96196 Refactor getMinMaxBoundForSBEType (#29327)
## Problem
There are a few problems in `getMinMaxBoundForSBEType`
* It combines retrieving min bound and retrieving max bound in one
function.
* When retrieving max bound for a type, the returned bound may not
always have the same type.
* The implementation is inconsistent with
[appendMaxForType](https://github.com/10gen/mongo/blob/cabb7b2a8b51564083507eeba575c9d9da1da170/src/mongo/bson/bsonobjbuilder.cpp#L114)
in BSONObjBuilder.

These problems make `isFullBracketInterval` difficult to address an
interval with two viable representation. For example, an interval for
predicate `{$gte: 100}` may be represented as
* [100, Inf.0]
* [100, "")

`isFullBracketInterval` is only capable of addressing the latter one by
assuming the types have to differ. This PR makes `isFullBracketInterval`
capable of addressing both by using the utilities factored out from
`getMinMaxBoundForSBEType`.

## Changes
1. Add utility function isVariableWidthType(TypeTag)
    1. It returns true for objects, arrays, etc. 
2. It returns false for numbers, dates, timestamps, null, undefined,
etc.
2. Add utility function getNextType(TypeTag)
    1. e.g. getNextType(NumberInt32) -> StringSmall
    2. e.g. getNextType(StringSmall) -> Object
3. Replace getMinMaxBoundForSBEType
1. getMinBound is mostly the same as `getMinMaxBoundForSBEType(type,
true /*isMin*/)`
2. getMaxBound can be simplified by determining if the given `type` has
variable width. If so, it can derive the bound by
`getMinBound(getNextType(type))`. This makes the switch case logic
shorter.
4. Improve `isFullBracketInterval` by using `isVariableWidthType`. Refer
to the comments in code for details.
5. Change implementation of `sameTypeClass` by comparing the sort order
instead. This improves the performance 3x in average, from 60ns to 20ns,
as it avoid invoking `appendMinForType` which may allocate memory for
some types.