SERVER-97077 Ban changeStream tests from running in sharding_csrs_continuous_config_stepdown (#29297)
Ban all changestream-related tests in suite
`sharding_csrs_continuous_config_stepdown` from running that are not
using the `ChangeStreamTest` fixture. The reason is that the tests that
are not using the `ChangeStreamTest` fixture do not handle
`ResumableChangeStreamError`s, so they can fail when there are
stepdowns.

Additionally we guard against new changestream-related tests without the
`ChangeStreamTest` fixture being added to the suite. This is achieved by
a command override that hooks into all aggregate commands with a
`$changeStream` pipeline and makes them fail with a descriptive error
message.