SERVER-96841 checkMetadataConsistency should not check the timeseriesBucketsMayHaveMixedSchemaData top-level catalog field (#29395)
This field is deprecated and can be inconsistent across shards, so
project it out of the `checkMetadataConsistency` aggregation to avoid
inconsistent values generating a `CollectionAuxiliaryMetadataMismatch`.

The test for `CollectionAuxiliaryMetadataMismatch` has been adapted to
use a fail point, as the deprecated timeseries flags are the only known
cases where the top-level metadata can be made inconsistent by MongoDB.
An inconsistency can also be simulated by modifying `_mdb_catalog` with
WiredTiger, but doing so would be significantly more involved.