SERVER-94181: streams dynamic checkpoint interval (#29347)
* Prior to this change, processors without windows checkpoint every 5
minutes, and processors with windows checkpoint every 60 minutes.
* Editing a stream processor with a window requires replaying from a
checkpoint before the window started. With the prior checkpoint
interval, this would mean editing a processor with tiny windows would
still require replaying up to one hour of data.
* In this PR we choose a checkpoint interval based on the size of the
last checkpoint. The interval chosen is between 5 minutes to 60 minutes.
If the last checkpoint was 0 bytes, the interval is 5 minutes. If the
last checkpoint was 100MB, the interval is 60 minutes.