SERVER-97306 Have scons hardlink files from bazel-out rather than copy (#29400)
We are using too much extra space by copying files from bazel-out to the
build directory. Replace the copy with a hardlink - we are hardlinking
rather than symlinking because SCons install was having issues with the
symlink. Some extra logic was needed to deal with edges cases around not
hardlinking across disks/volumes.