SERVER-96995 (QueryTester) Communicate metadata to users about failing queries (#29119)
This PR integrates the feature extractor to communicate metadata about
failing queries after running tests through a mongotest invocation. As
the feature extractor takes more time to process failing queries
compared to the text-based diff, this is provided as an option `-v`
(verbose) that users can specify for local testing and debugging. It
won't be run by default on patch builds.

Note that this assumes that the user has the feature-extractor repo
checked out locally in their home directory. I'm open to suggestions for
more streamlined, alternative approaches.

The metadata is displayed after the failure summary:

![metadata](https://github.com/user-attachments/assets/a4807f6e-50d7-4150-b405-9d9a898197f8)

Note that this was ported over from
[DEVPROD-12318](https://github.com/10gen/QueryTester/pull/49)