SERVER-93874 Only use time-series control.max bucket metadata when the date is after unix epoch (#29352)
The root cause is query code not accounting for
[SERVER-67816](https://jira.mongodb.org/browse/SERVER-67816), which can
cause wrong results when we try to use control.max when the bucket has
data <1970. This change makes two fixes:

Don't use bucket-only plans (plans that only access the bucket metadata)
if we need the control.max field for time.
Block processing will check the date before using the bucket control.max
for time.