SERVER-94946 Add prove-server-programmability target (#29261)
This change adds a `prove-server-programmability` SCons target that can
be used to build and run all server programmability unit tests, with the
build system keeping of track of which ones need to be rebuilt and
reran. Tests are included in that target by passing
`PROVE_ALIASES=["server-programmability"]` to the `CppUnitTest`
function.

This change also tags the few existing bazel unit tests owned by SP with
the `server-programmability` tag, which serves the same purpose in
bazel.

Some tests have been split into two as they contained both SP-owned and
non-SP-owned tests that were fairly easy to disentangle while
maintaining logical boundaries. `util_test` also contains a mix of SP
and non-SP tests, but it's mostly SP tests and there's not a clear
logical way to split it up, so I've left it as is for now.