SERVER-94947 Add Server Programmability smoke test suite (#29379)
This change adds a smoke test suite for the server programmability team.

More generally, it adds a run_smoke_tests.py buildscript that refers to
a very simple yaml config file that defines a smoke test suite. The
script executes a series of resmoke.py commands to actually run the
tests.

We ultimately want the smoke tests to run in the commit queue, but this
will happen separately.

---------

Co-authored-by: Alex Li <98560274+lia394126@users.noreply.github.com>