SERVER-97088 Use histograms to estimate cardinality of sargable MatchExpressions (#29378)
The idea of this patch is to make use of histograms available when we
encounter a `ComparisonMatchExpression`, since we know that expression
can be easily translated to an equivalent interval form. Once we have
the interval, we can invoke histogram CE the same way we do when we
encounter an IndexScan.