SERVER-94855 Support non-type-bracketed mixed-type interval histogram CE estimation (#29121)
Supports a wider range of intervals which are not type-bracketed but
still estimable when both bounds are estimable types. The strategy is to
divide the interval into several sub-intervals, each of which is
type-bracketed and satisfies sameTypeBracketInterval(). Then, we can
estimate each sub-interval using either histograms or type counts.

Example:
An interval [123, true] can be divided into the following sub-intervals:
1. [123, inf.0]: This sub-interval can be estimated using a histogram.
2. ["", {}): This represents a full-bracket interval for strings, which
can be estimated using type counts.
3. Other full-bracket intervals for objects, arrays, etc., which can
also be estimated using type counts.
4. [false, true]: This sub-interval can be estimated using the counts of
false and true values.