SERVER-90874: Add --directoryPerDb and --wiredTigerDirectoryForIndexes to the config fuzzer (#29389)
# Issue

[SERVER-90874](https://jira.mongodb.org/browse/SERVER-90874)

# Description

We add these additional configuration options to the config fuzzer in
order to improve the robustness of our tests to changes in this
configuration. There are a few important differences between those
configurations at the configurations currently supported by the config
fuzzer:
- These are not `setParam`s, so they have to be applied to the mongod
binary differently compared with the other fuzzed configs
- These are *not* to be exercised by jstests (see discussion in #24928,
and SERVER-92894)
- These configurations must be present when re-starting a mongod
instance if it was initially launched with them.

---------

Co-authored-by: Matt Kneiser <matt.kneiser@mongodb.com>