SERVER-97576: Use full coll path when considering --drop --load optimization in QueryTester (#29524)
Instead of using the raw collection string in the test file, which is
based on a relative path, this now uses the a combination of the
collection name and the full path of the data source. This ensures that
distinct data sources that are loaded with the same filename and
collection name in different tests within different directories will not
be considered equivalent collections for the purpose of this
optimization.