SERVER-95901 Introduce $listMqlEntities stage (#28900)
This patch introduces a new DocumentSource `$listMqlEntities` which
initially only supports returning the set of available DocumentSources
which the parser accepts. This allows us to write jstests and verify
that a test-case exists for each stage. We replace the command auth
tests to use this mechanism instead of relying on serverStatus counters
to determine the set of stages. This stage is extensible and can be
extended to support other MQL entities like agg expressions, Match
expressions, etc.