SERVER-96412 Fix and clarify shard filtering contract for $search (#28701)
This fixes an issue where an operation which needs to target two
different collections can end up mistakenly adding a shard filter
operation for a PlanExecutor which is not intended to have one.

Specifically, we observed this case in an aggregation pipeline like
`db.x.aggregate([{$unionWith: {coll: "y", pipeline: [{$search: {}}]}}])`
executing in a 1-shard sharded cluster, with both collections ("x" and
"y") sharded. In this case the $unionWith targeting logic will try to
directly execute a sub-pipeline locally on the same shard, rather than
using the network protocol and sending a request to itself. The key
difference here is whether there is a new OpCtx created. A directly
executed sub-pipeline will re-use the same OpCtx. Sending a request over
the network to the same host will result in a new OpCtx for this
sub-command. For this bug, the re-used OpCtx gets passed to an
invocation of get_executor's logic to construct a PlanExecutor over
collection "y". The logic then breaks down when the OpCtx shows _some_
shard version (for "x"), but none for collection "y". In this case the
lack of a version for "y" was intentional, and the expectation was to
not have a shard filter applied.

There are a couple key conditions for reproducing this which were not
obvious at first:
* The shard key must _not_ be _id, but can be anything else. There is a
special codepath for express queries which avoids this problem, likely
due to the special case described in SERVER-78071. This root cause was
only confirmed with code inspection though. I do know that the issue
doesn't reproduce with an _id shard key.
* This is specific to $search and idLookup, and does not affect a
similar $unionWith pipeline which uses a more normal $cursor stage at
the front. This is because for a "normal" $cursor codepath, the shard
filtering is expected in the cursor, and the get_executor.cpp
sub-routine is invoked while the OpCtx has two namespaces set up with
shard versions - so everything works out.