SERVER-97656 Wait for operations to be queued before sleeping in ingress admission control js tests (#29413)
The test was failing since we couldn't make the javascript precisely
start waiting until time started to get accumulated in
totalTimeQueueMicros in the admission context.

Using curop makes the wait reliable

We also do have a way in the C++ to detect when the ticket is waiting
after time accumulation has started.

---------

Co-authored-by: Matt Broadstone <mbroadst@mongodb.com>