SERVER-97673 Use aliased collection name instead of full collection line in failed query feature extractor script (#29644)
For a collection line in a `.test` file that looks like
`<real_name>.coll as <alias_name>`, we will use `<alias_name>` if it
exists. If not, we will use `<real_name>`.

[Previously](https://github.com/10gen/mongo/pull/29119/), we were seeing
`PlanStage::EOF` because we were incorrectly extracting the line
`<real_name> as <alias_name>` as the collection name. Now that we are
referencing the right collection name, we are seeing correct plan stage
usage.

<img width="762" alt="Screenshot 2024-11-26 at 5 58 00 PM"
src="https://github.com/user-attachments/assets/2139d4f0-a470-4bf0-911a-d8d8da35bef2">