SERVER-97694 Deduplicate link flags when linking statically (#29670)
Previously the link line would be duplicated for every layer of depth in
the graph, which was causing bad performance on windows. Perform the
deduplification manually.

Removing this increases the performance of the link by around 60% on
Windows.

Ex:

original link line count on util_test:
9326

new link line count on util_test:
741