SERVER-97612 Make signal processing thread start before signal testing thread (#29599)
When we added a thread to test EINTR resilience (hidden behind a
failpoint), we made that thread start before the signal handler thread.
However, it needs to start after the signal handler, or at least after
the signal mask is configured on the initial thread. Otherwise, the
testing thread can respond to signals. This can cause problems; when the
signal handler thread handles SIGTERM, mongod exits with code 0, but
when the testing thread handles SIGTERM, mongod exits with a nonzero
exit code, confusing the test infrastructure.

This change makes the testing thread start after the signal processing
thread.