SERVER-76390 Quickly determine all indexes affected by writes (#27568)
Extend the collection/index catalog so that it maintains a new helper
object (`IndexUpdateIdentifier`) that keeps track of the fields indexed
by the indexes of the collection.
This `IndexUpdateIdentifier` object is used to quickly determine which
indexes of the collection may be affected by a write operation.
The `IndexUpdateIdentifier` object is maintained separately per
collection. The instances live in `index_catalog_impl.h`.

The `IndexUpdateIdentifier` is updated whenever any index is added to or
deleted from the catalog entry, i.e. whenever any of the `_readyIndexes`
or `_buildingIndexes` containers gets modified. It is intentionally not
updated when `_frozenIndexes` is modified, because frozen indexes do not
participate in write operations.

To update the `IndexUpdateIdentifier` after indexes have been added or
removed, all indexes present in the members `_readyIndexes` and
`_buildingIndexes` are iterated (in this order). For each index
contained in any of these containers, we collect the already existing
`UpdateIndexData` helper object, and then add it to the
`IndexUpdateIdentifier` with the position of the current index inside
the `_readyIndexes` / `_buildingIndexes` continuum.
For example, if there are three indexes on a collection, `{ "a": 1 }`
and `{ "b": 1 }`, and `{ "a": 1, "c": 1 }`, the `IndexUpdateIdentifier`
will have the following bits set for the fields used by the indexes:
- a: [0, 2]
- b: [1]
- c: [2]

The update code path (`IndexCatalogImpl::updateRecord`), when given an
`opDiff`, makes use of the new `IndexUpdateIdentifier` to quickly figure
out which indexes may be affected by the update. For that it calls into
the `IndexUpdateIdentifier::determineAffectedIndexes` method. For
example, if the `opDiff` contains a modification to field "a", we probe
the `IndexUpdateIdentifier` for the field name "a" and it will return
indexes [0, 2] to us. This means we need to update index 0 and index 2
in the `_readyIndexes` / `_buildingIndexes` continuum.
The `IndexUpdateIdentifier` also contains a separate bit set for
wildcard indexes (i.e. `$**`). For every wildcard index, a bit is set in
this vector.

The change already removes the `_indexedPaths` member from the
`CollectionQueryInfo` class. `_indexedPaths` was only populated, but
never read back. The instance variable `_keysComputed` is also removed,
because it was also unused.