SERVER-97071 Ensure stability of resumeTokens across versions (#29153)
SERVER-88142 introduced an additional subfield `type` inside the
`operationDescription` of "create" changestream events, in case that the
changestream was opened with the "showExpandedEvents" flag set.

It turned out that extending the `operationDescription` field with
additional subfields is not a good idea, because the full
`operationDescription` value is used to generate the resumeToken for the
event.

That means the resumeToken for a "create" event would be different
depending on whether the new "type" field is produced or not. This could
be a problem during upgrades.

This PR reverts the changes from SERVER-88142, so that the
`operationDescription` values for "create" changestream events will
remain the same when upgrading from older versions to 8.1. The
resumeTokens will remain fully compatible.
The previous `type` field will now be exposed in a new top-level field
named `nsType` if the changestream was opened with the
"showExpandedEvents" flag. The `nsType` field will contain either
"collection", "timeseries" or "view".