SERVER-95748 createRegex() in bsoncolumn_builder_fuzzer.cpp test code needs work (#29245)
**Changes:**
- `createRegex` function now takes in both pattern and option arguments
- When we are generating a `Regex`, we generate both a fuzzed pattern
and options string. We make it so that the options and pattern string
are distinct from eachother and we advance the pointer that represents
the fuzzed data we are creating the fuzzed `BSONObj` up to` len + 1` so
we avoid using the null terminator for the next part of the buffer.
- Made it so that we can also "generate a string" in the case that there
is end == ptr (empty string).
- `generateBufNoNuls` ~~now return a `len` rather than a bool for if we
are able to successfully generate a no-nul string buffer (still enables
us to generate empty strings), which enables us to advance the pointer
only up to the point we actually used the memory~~

---------

Co-authored-by: Shin Yee Tan <shinyee.tan@mongodb.com>