SERVER-95926 Create a ShardServerCatalogCacheLoader interface for its implementations (#28729)
Rework the class hierarchies of `CatalogCacheLoader` and turn
`ShardServerCatalogCacheLoader` to be an interface.
`ConfigServerCatalogCacheLoader` and `ShardServerCatalogCacheLoader`
independently extend `CatalogCacheLoader`. Some filtering metadata
exclusive methods have been moved to `ShardServerCatalogCacheLoader`, to
simplify code and remove some `dynamic_cast` warts.

In the future, with SPM-3729 (authoritative shards), the
`ShardServerCatalogCacheLoader` branch will be presumably deleted and
the `ConfigServerCatalogCacheLoader` branch simplified to be simply
`CatalogCacheLoader`.