SERVER-96792: set 'score' metadata field alongside 'searchScore'/'vectorSearchScore'/'textScore' (#29124)
[SERVER-96792](https://jira.mongodb.org/browse/SERVER-96792)

For `$rankFusion` and `$scoreFusion`, we plan on making their score
accessible by the metadata field name `score` (instead of
`rankFusionScore` / `scoreFusionScore`).

In this PR, we are migrating the other specialized score metadata field
score types over to the same generalized `score` metadata (but do not
remove any existing metadata for backwards compatibility). This change
will allow the scores of the input pipelines to `$scoreFusion` to be
fetched agnostically, and will be needed to compute the `scoreDetails`
agnostically regardless of the type of input pipeline to `$rankFusion `/
`$scoreFusion`