SERVER-96733 Failing to validate a timeseries bucket leads to not logging timeseries errors related to containing mixed schema data (#29815)
Even if the bucketStatus is not ok from the
`_validateTimeSeriesBucketRecord`, we still do validation on if there is
mixed schema data in the bucket.
Test change was because now we have two errors during validation (the
bucket status not being ok and having a malformed min/max) instead of 1.
I don't think we need to add more testing to ensure that both 6698300
and 846990[0-2] gets logged together when there are cases of these two
instances.