SERVER-97845 Work around pipeline_validation.js failures (#29824)
This is a temporary fix to quiet build failures.

My best theory so far is that there's a timing issue with a sharding
catalog refresh thread which causes the query to sometimes fail and be
retried, and when it fails the first time it must incorrectly assume
it's sharded or something? Not a whole lot to go off of so far.
[SERVER-97846](https://jira.mongodb.org/browse/SERVER-97846) tracks the
larger fix.