SERVER-97819 Use RAIIServerParameterControllerForTest for server parameter initialization in tests (#29882)
There is a potential data race involving the gMultitenancySupport
variable. Some tests change the value on the beginning. The data race
happens when getOrGenerateSampleId is called from a different thread
while the test is running.
The solutions is to set the value of gMultitenancySupport for some tests
before the test is initialized.