SERVER-97103: Allow sorting by {$meta: "score"} (#29840)
The metadata field "score" should now be present everywhere
"vectorSearchScore" and "searchScore" (as well as "textScore') are. This
PR simply makes "score" a valid metadata field to sort on (as only a
subset of metadata fields are allowed).

We already had robust testing for sorting on "vectorSearchScore" and
"searchScore", so I just modified those same tests to also sort on
"score" too.