SERVER-87743 Make ReadWriteConcernDefaults managed on a per-service basis (#26511)
ReadWriteConcernDefaults have different behavior on get/set depending on
the cluster role using the defaults. The router role always refreshes
defaults from the config server and only caches in-memory locally. The
shard-role does not manage the defaults at all in a sharded cluster,
unless it is a config-server, in which case it also keeps a local
in-memory cache of the defaults and persists them locally.
This PR refactors the management of ReadWriteConcernDefaults to keep
role-aware caches and accessors to keep the different behavior preserved
even across different roles in the same process.