SERVER-97904: Copy BSON before operation active check (#29509)
[BF-35228](https://jira.mongodb.org/browse/BF-35528)

The aggregation pipeline:

`[{"$currentOp":{"idleConnections":false,"allUsers":true,"truncateOps":false}},{"$match":{"active":false}}]`
which by definition should return zero results, had a case where it
returned one, as the `$currentOp` stage returned a result of the
`OplogWriter` being inactive, when only active results were requested.

This is because in this case, the `OplogWriter` turned from active to
inactive between the time its status was checked (old L113) and the time
it was converted to BSON to be included in the results (old L120).

This PR provides a simple solution where first the Client/operation is
converted to BSON, and then the active status is checked, so that
nothing can be concurrently updated between checking and returning the
results.

This fix is of course not comprehensive to all possible bugs of this
type (where some relevant state of the operation changes after it is
checked, but before it is returned). However, while investigating this
BF and implementing this PR, I determined that generalizing this
strategy is not so easy given the current code structure. I wrote and
backlogged [SERVER-97558](https://jira.mongodb.org/browse/SERVER-97558)
to think proactively about other possible incorrect query results of
this type, while ending this BF in a timely manner.