SERVER-97365 Add smoke tests to commit queue (#29490)
This change adds the smoke tests (currently only defined for server
programmability) to the commit queue. It includes a minor refactor of
the recently-added `run_smoke_tests.py` script so that it can (1)
coalesce multiple smoke test suites into a single run; and (2)
automatically discover and run all smoke test suites by default (i.e.,
if no suites are passed).

A new script `evergreen_gen_smoke_test_tasks.py` runs in the commit
queue, it uses the logic from `run_smoke_tests.py` to discover and
coalesce the smoke tests and generate a task for each suite. The simpler
alternative would be to directly run `run_smoke_tests.py` in the commit
queue, but this wouldn't parallelize the suites and wouldn't use all of
the infrastructure we already have for running resmoke suites in
evergreen.