SERVER-97629 Replace invariant with tassert when yielding (#29603)
In `plan_yield_policy.cpp` it seems possible that under some condition
we either do not have a `yieldable` pointer to a yieldable object, which
can trigger an invariant failure.
The issue was surfaced by
[AF-1175](https://jira.mongodb.org/browse/AF-1175).
The exact reason for the yieldable object not being present is unknown,
but we replace the invariant with a tassert and provide additional
information in case it is triggered.