SERVER-96106 Support EINTR/SIGUSR2 testing via config fuzzer (#29737)
This change adds two new fuzzing-related options:

1. In the config fuzzer, we can now enable the recently-added failpoint
enableSignalTesting at startup time in both `mongod` and `mongos`. This
failpoint periodically chooses a victim thread and sends it a signal,
which can interrupt running system calls. This is intended to test
resilience to EINTR results from syscalls, which has been a (very rare)
source of bugs in the past.
2. It adds a mode for `--fuzzRuntimeStress` that will periodically send
`SIGUSR2` to a victim process started by the fixture. This signal
requests a stacktrace from all threads in the process. This has the
effect of stopping all threads, making them dump their stacks, and
continuing on. This tests the robustness of our stack tracing system,
which has been a source of bugs in the past as well.