SERVER-97468 Fully disable trackUnshardedCollectionsUponCreation (#29452)
The feature flag trackUnshardedCollectionsUponCreation is currently
disabled in production. The code behind the feature flag still requires
to see the completion of SPM-3680, which won't make it for v9.0.
Meanwhile, the flag is known to be causing a possible deadlock
(SERVER-95544).

The ticket proposes to:
- Fully disable the feature flag, this is the first step to prevent the
deadlock in a future ticket
- Still maintain the coverage by keep supporting tracking upon creation
on specific suites via a test command
**The suites are necessary** as they are critical to maintain an
extensive coverage for queries on collection that are tracked. As
opposed to the feature flag, the suites only use a test command which
won't be supported in production.

The ticket mainly added a tag and a global variable to the suites where
collection are tracked, to be able to exclude test or test cases (where
before the feature flag was checked)