SERVER-97800 Include counts for Null and Nothing in [null, null] interval estimation (#29929)
This commit fixes the bug that values missing from a document are not
reflected in the histogram estimation result. For instance, the estimate
of predicates such as `{a:null}` reflects only the documents where `a`
is explicitly null but not the documents where field `a` is missing
altogether. This is because in histogram generation, documents missing
the field increments the counts for `TypeTags::Nothing` rather than
`TypeTags::Null`.

The fix is to add the counts from both Null and Nothing during the type
counts estimation when the query interval is `[null, null]`.