SERVER-97339: Implement serialization for $score with generic $setMetadata internal operator (#29911)
The primary purpose of this PR is implementing serialization of the
$score stage, in order to set us up for testing $score end-to-end in
sharded environments.

Since $score is actually an alias for the use of the
`SetMetadataTransformation`, the core of this PR is actually
implementing an internal `$setMetadata` operator (using
`DocumentSourceSingleDocumentTransformation` +
`SetMetadataTransformation`). This PR also enables the
`SetMetadataTransformation` to set any metadata type, rather than just
the score, and tests the query shape generation of the `$setMetadata`
and `$score` stages.