SERVER-98076 Create library compile task for unit tests to depend on (#30039)
Previously the 4 unit test tasks scheduled in parallel would end up
tearing through the cached part of the build, and then all getting to
the same spot, compiling each C++ file 4x since the artifact wouldn't be
uploaded to cache in time for the others to retrieve it. This put 4x
load on the remote execution cluster.

Share remote cache of core libraries between unit test groups by
creating a shared pre-unit test target for them to depend on.