SERVER-98044 Remove pin_code_segments_on_startup.js from smoke suite (#30015)
This test depends on the host's configured ulimits so it can fail
locally if those aren't sufficient. It seems like maybe the size of our
code segments has recently passed the default memlock limit on our
virtual workstations, making this test start failing locally despite
passing in evergreen.

Another option would be to allow tagging to select smoke tests, e.g., we
could exclude tests tagged with `requires_increased_memlock_limits`
(which only applies to this test). It doesn't seem necessary at this
point, but that may be a better option if we find we want more ways to
control the suite over time.