SERVER-96381 Kill backup cursor from previous FCBIS attempt in fcbis_restart_node_during_sync_with_pre_images.js (#29554)
This test has multiple FCBIS attempts and sometimes can fail to set the
appropriate failpoints if a backup cursor from a previous attempt is
still open. This is because we use a fire and forget command to kill the
cursor in the server. While this isn't much of a problem in prod because
we have a cursor timeout and don't expect customers to run back to back
FBCIS attempts within that window, we should still try to fix the test.
This test now manually runs killCursors with the appropriate backup
cursor ID if we encounter error 5973000.