SERVER-95448 Split getShardIdsForRange improving the visibility of pe… (#29028)
…rformace degradation risk for chunks.

This call of getShardIdsForRange is copying the entire chunk map which
can contain thousands or even millions of chunks. This can take a very
long time, and so having this as part of the getShardIdsForRange
function is very dangerous as it can introduce huge performance hits if
used incorrectly.

Also the set of ChunkRanges is used as parameter in many methods and
functions although it's not used in the production code currently,
except _getTargetingInfoForQuery which is used for analyzeShardKey and
tests. All such places are cleaned up removing usage of ChunkRanges.

Changes: 
* The existing functions are renamed to getShardIdsAndChunks* to make it
clear, that it also handles chunks.
* A new function getShardIds* is created, which does not allow to
provide chunkRanges, to avoid a risk to fill it by mistake.
* All calls where chunks not needed remain to use getShardIds* without a
chunk parameter.
* targetAllShards is cleaned up removing chunks parameter as it's not
used.
* Only analyzeShardKey and tests are providing the chunk parameter.
* The ChunkRanges parameter is removed from targetUpdate and
targetDelete as it's not used.
* The ChunkRanges parameter is removed from
CollectionRoutingInfoTargeter::_targetQuery as it's not used.
* The unit tests related to ChunkRange are moved from targeter test
collection_routing_info_targeter_test.cpp to
shard_key_pattern_query_util_chunk_ranges_test.cpp to test directly
against getShardIdsAndChunksForCanonicalQuery.