SERVER-97874 Adjust code-lockdown thresholds at the VP level (#29870)
Recently, we've come very close to an server wide code-lockdown without
the VP level quota protecting us.
This lowers the quotas based on historical data to something that would
have fired earlier, although maybe still not early enough in all cases.

Based on recent numbers, this would put DevProd into a code-lockdown,
but the intention with rolling this out would not be to immediately
enforce that given that we have only started tracking rolled up stats
for DevProd in the last few weeks. We'll co-ordinate the one-time
rollout expectations there separately.