SERVER-98297 Run test_simple_compares.py and test_simple_run.py in serial (#30175)
Ensure `simple_run` and `simple_compare` are run in serial so that a
testfile isn't processed with `--mode run` at the same time that it's
processed with `--mode compare`. This can lead to race conditions
between deleting a `.actual` file (in `--mode run`) and reading it (in
`--mode compare`).

Fix for [BF-35936](https://jira.mongodb.org/browse/BF-35936).