SERVER-97939 Reduce number of temporary objects (#29943)
Some code parts use `std::vector::emplace_back(...)` in a non-optimal
way.

Instead of passing the arguments for constructing the object in-place,
we are passing a temporary object. The temporary object will then be
std::moved into the vector by the `emplace_back` call. This can be
improved by just passing the arguments to construct the object in-place.