SERVER-96460 Validate against invalid PQS hints (#29711)
### Summary
This PR adds extra validation for natural and key-pattern index hints.
It also removes invalid query settings index hints on cluster parameter
refresh.

### Detailed explanation
The following validations were added for index hints on new
setQuerySettings commands:
- index key-pattern indexes can only contain numbers as their values
(eg. `{a: "1"}` will be invalid).
- natural hints can only contain one fields (eg.`{$natural: -1, a: 1}`
will be invalid).
- empty key-pattern indexes are not allowed (eg. `{}` will be invalid).

On cluster parameter refresh, we also prune away all indexes which are
not valid by the above metrics, in order to ensure all query settings
will be valid until next release. The aim is to remove this extra step
once we are sure not such invalid index hints are present.

### Notes
This PR also adds `query_settings_utils_test.cpp` to a unit test target,
as it was not being run together with the other tests. Moreover, it also
adds missing imports and removes changes to the test that occurred prior
to this PR.

---------

Co-authored-by: Catalin Sumanaru <catalin.sumanaru@mongodb.com>