SERVER-96154: Validate $rankFusion input pipeline names (#30174)
This PR validates that the $rankFusion input pipeline names are valid
field paths (for example, do not start with a "$" or contain "."
characters).

I re-used the `FieldPath::uassertValidFieldName` utility to do so;
however, the utility left a misleading error message suggesting that a
user _could_ have a field name with dots or dollars if you used
"$getField" or "$setField", which is not true when validating $facet or
$rankFusion pipeline names. I modified the utility function to omit that
suggestion when invalid in those cases.