SERVER-98325 Deduplicate failed query ids by using a set instead of a vector (#30207)
We can get duplicate failed query ids when using a `std::vector` when
the `git diff` output for a failing query is large and spans multiple
chunks. This led to misreporting of the number of failed queries when
running QueryTester in compare mode with verbose logging (`-v`). Using a
`std::set` resolves these issues.